Add test helpers for unstaged/uncommitted changes
In order to better test detection of unstaged and uncommitted
changes, add base test methods to create changes and not stage or
commit them, ensuring they can result in an unstaged diff or staged
but uncommitted edits. To ensure that these work, use them within
the simple change creation method to perform the actual file edit.

Change-Id: Ib698d0057a404f073490d1683a8eef8d0c143122