Don't test rebasing with unstaged changes
For safety, attempts to push a commit with unstaged or uncommitted
changes in the worktree will be caught and an error reported, rather
than leaving it up to ``git rebase`` to spot them. This addresses a
situation where users enabling "rebase.autostash" would otherwise
experience data loss when the test rebase is subsequently reset.

Change-Id: Iabb8387c9db59a7d02ebfd43b688e7bb93d3159f
Task: #38921