Increase test timeout to 5 minutes
A 2-minute global test timeout was added to the base test class back
in 2014, to prevent tests from hanging until the job timeout is hit.
This has served us well, but with updates to Gerrit and increasing
complexity of our tests, we often get a slew of test timeouts on
some CI nodes (ad on my workstation for that matter!). Increase this
to 5 minutes, which should still serve the intended purpose while
providing more time for tests to complete on less-performant
systems.

Change-Id: I9673ce3cf18f9e4c96fd7e9f96f5ddce2ef6d957