Install PBR explicitly
We do this because easy_install on python3.5 is failing to find pbr. I
believe this is happening due to pypi's deprecation of non SNI tls
connections. They report that the version of 3.5 on xenial should be
able to do SNI but easy_install at least doesn't seem to do so.

Note we restrict this to python>=3 and <=3.5 as python 2.7 had
backported fixes. Also, this really only helps with unittesting via tox
as we instruct tox to explicitly install the requirements files before
creating/installing the package itself. This is the reason we stick this
in test-requirements.txt rather than requirements.txt.

Change-Id: Iac2c1f4f575de250ed0dba22d33b017322fd7562