Clean up package metadata
The author-email and python-requires options are deprecated by
SetupTools in favor of author_email and setup_requires and trigger
warnings now, so update them. The license_files option (since its
introduction) automatically includes LICENSE and AUTHORS if not
specified, so was always redundant and unnecessary. The warnerrors
option for PBR was for Sphinx integration, which we don't use any
longer, and is ignored by modern PBR versions anyway so clean it up
too.

Change-Id: Ic5d1e17dd4926eea4e5b6fff47295fd8e9eb452d